vn_create is changed to pass a vnode via *vpp with VOP_CREATE if the file was found. All file systems need to VN_RELE if *vpp is set when VOP_CREATE is called. This was done before in vn_create. This tells the underlying file system that the file "name" already exists and its vnode is given in *vpp. NFS and others should be modified to take advantage of this.
This checkin also inserts the VOP_VNODE_CHANGE calls into xfs and efs
when a create hits an existing file and that file is truncated.
This simplifies CFS (ds side) when a file is created.

Also, the xfs_create code is cleaned up and simplified by taking
advantage of the *vpp passed in the vn_create. The creat-open
aim test is now 3 times faster.