Patch from: William L Jones It use anonymous dcache entries like nfssd. It use dentry_open from fs/open.c to do the mundane work of creating a file entry.
I hope this will meke the code cleaner and easier to maintain and I think it  gets 
rid of some potentially problems that can come about with the way nfsd use 
dcache entries.  It really  needs will formed cache entries for files 
and directories or for them to be anonymous so it can ignore them.