dmapi - Execution of an offline script or binary fails.
If a user thread is trying to execute the file that is offline then the HSM
won't get write access when it attempts invisible I/O to bring it online
because the user thread has already denied write access...but that thread is
waiting for us to write the file....  So add a callout from open_exec() to
give DMAPI an early notice that the file must be online.
Add linvfs_open_exec, and put it linvfs_file_operations.