Rework prohibited_mr_events hint for xfs_dm_set_region xfs_dm_punch_hole(): no longer needs to check VN_MAPPED, because it's okay to punch out mapped files now.  Active executables are protected by the call to get_write_access().
xfs_dm_set_region(): This is often called by DMF to simply remove the
READ event bit without changing any other bits, and this is behavior
that DMF relies on, so change the code to do it this way.

prohibited_mr_events(): Change this to search all writable maps.  Fix
for the 2.6.11 kernel, too.