write barrier support
Issue all log sync operations as ordered writes.  In addition flush the
disk cache on fsync if the sync cached operation didn't sync the log to
disk (this requires some additional bookeping in the transaction and log
code).
If the device doesn't claim to support barriers, the filesystem has an
extern log volume or the trial superblock write with barriers enabled
failed we disable barriers and print a warning.  We should probably fail
the mount completely, but that could lead to nasty boot failures for
the root filesystem.

Not enabled by default yet, needs more destructive testing first.