Initial pass at going directly-to-bio on the buffered IO path.
This allows us to submit much larger I/Os instead of sending down lots
of small buffer_heads.  To do this we need to have a rather complicated
I/O submission and completion tracking infrastructure.  Part of the latter
has been merged already a long time ago for direct I/O support.
Part of the problem is that we need to track sub-pagesize regions and for
that we still need buffer_heads for the time beeing.  Long-term I hope
we can move to better data strucutures and/or maybe move this to fs/mpage.c
instead of having it in XFS.

Original patch from Nathan Scott with various updates from David Chinner
and Christoph Hellwig.