clean up the xfs_offset_to_map interface
Currently we pass a struct page and a relative offset into that page around, and returns the current xfs_iomap_t if the block at the specified offset
fits into it, or a NULL pointer otherwise.  This patch passed the full 64bit
offset into the inode that all callers have anyway, and changes the return
value to a simple boolean.  Also the function gets a more descriptive name:
xfs_iomap_valid.