Fix xfs_splice_write() so appended data gets to disk.
xfs_splice_write() failed to update the on disk inode size when
extending the so when the file was closed the range extended by
splice was truncated off. Hence any region of a file written
to by splice would end up as a hole full of zeros.
Merge of xfs-linux-melb:xfs-kern:26920a by kenmcd.

  Update xfs inode size if xfs_splice_write is writing beyond the
  end of the current file.