Prevent a deadlock when xfslogd unpins inodes.
The previous fixes for the use after free in xfs_iunpin left a nasty
log deadlock when xfslogd unpinned the inode and dropped the last
reference to the inode. the ->clear_inode() method can issue
transactions, and if the log was full, the transaction could push on
the log and get stuck trying to push the inode it was currently
unpinning.

To fix this, we provide xfs_iunpin a guarantee that it will always
have a valid xfs_inode <-> linux inode link or a particular flag
will be set on the inode. We then use log forces during lookup to
ensure transactions are completed before we recycle the inode.
This ensures that xfs_iunpin will never use the linux inode after it
is being freed, and any lookup on an inode on the reclaim list will
wait until it is safe to attach a new linux inode to the xfs inode.
Merge of xfs-linux-melb:xfs-kern:27359a by kenmcd.

  If we find a reclaimable inode, make sure it is not pinned before we
  reuse it so we don't get inode unpinning running incorrectly on a
  new linux inode. Also ensure that we don't remove unlinked reclaimable
  inodes from the reclaim list which would cause them to leak.