Prevent buffer overrun in cmn_err().
The message buffer used by cmn_err() is only 256 bytes and some CXFS
messages were exceeding this length.  Since we were using vsprintf()
and not checking for buffer overruns we were clobbering memory beyond
the buffer.  The size of the buffer has been increased to 1024 bytes
so we can capture these larger messages and we are now using vsnprintf()
to prevent overrunning the buffer size.
Merge of xfs-linux-melb:xfs-kern:27561a by kenmcd.

  Prevent buffer overrun in cmn_err() by using vsnprintf() and increase
  buffer size to 1024 bytes.