Current usage of buftarg flags is incorrect.
The {test,set,clear}_bit() operations take a bit index for
the bit to operate on. The XBT_* flags are defined as bit
fields which is incorrect, not to mention the way the bit
fields are enumerated is broken too. This was only working
by chance.

Fix the definitions of the flags and make the code using
them use the {test,set,clear}_bit() operations correctly.
Merge of xfs-linux-melb:xfs-kern:27565a by kenmcd.

  Fix enumeration of xfs_buftarg_flags_t to declare bit indexes
  rather than bit fields as required by test_bit() and friends.