Clear unwritten buffer flag on partial page invalidation
If we have a page with unwritten extents on it, and we do a
partial page truncation, we fail to clear the private unwritten
buffer flag on buffers that are now past EOF. If we then
extend the file into those blocks via a write, we will make
the wrong allocation decision when writing the new data
to disk because of the incorrect unwritten flag on the buffer.

Hence we need to clear the unwritten flag in xfs_vm_invalidatepage
to ensure that the state of the buffers on the page beyond EOF
are in the correct state in case they are reused.
Merge of xfs-linux-melb:xfs-kern:27804a by kenmcd.

  Clear private buffer flags in xfs_vm_invalidatepage to ensure
  buffers are consistent for reuse after a partial page truncation.