Make growfs work for amounts greater than 2TB
The free block modification code has a 32bit interface, limiting
the size the filesystem can be grown even on 64 bit machines.
On 32 bit machines, there are other 32bit variables in transaction
structures and interfaces that need to be expanded to allow this
to work.
Merge of xfs-linux-melb:xfs-kern:27894a by kenmcd.

  Change free block modifications to use 64 bit type casts.