Ensure a frozen filesystem has a clean log before writing the dummy record.
The current Linux XFS freeze code is a mess. We flush the metadata
buffers out while we are still allowing new transactions to start
and then fail to flush the dirty buffers back out before writing
the unmount and dummy records to the log.

This leads to problems when the frozen filesystem is used for
snapshots - we do log recovery on a readonly image and often
it appears that the log image in the snapshot is not correct.
Hence we end up with hangs, oops and mount failures when trying
to mount a snapshot image that has been created when the filesystem
has not been correctly frozen.

To fix this, we need to move th metadata flush to after we wait for all
current transactions to complete in teh second stage of the freeze.
This means that when we write the final log records, the log should
be clean and recovery should never occur on a snapshot image created
from a frozen filesystem.
Merge of xfs-linux-melb:xfs-kern:28010a by kenmcd.

  Add definition of SYNC_DIO_WAIT for telling the sync code
  to wait for direct I/O to complete.