Fix callers of xfs_iozero() to zero the correct range.
The problem is the two callers of xfs_iozero() are rounding out the range
to be zeroed to the end of a fsb and in some cases this extends past the new
eof.  The call to commit_write() in xfs_iozero() will cause the Linux inode's
file size to be set too high.
Merge of xfs-linux-melb:xfs-kern:28013a by kenmcd.

  Fix callers of xfs_iozero() to zero the correct range.