Fix to prevent the notorious 'NULL files' problem after a crash.
The problem that has been addressed is that of synchronising updates of
the file size with writes that extend a file.  Without the fix the update
of a file's size, as a result of a write beyond eof, is independent of
when the cached data is flushed to disk.  Often the file size update would
be written to the filesystem log before the data is flushed to disk.  When
a system crashes between these two events and the filesystem log is
replayed on mount the file's size will be set but since the contents never
made it to disk the file is full of holes.  If some of the cached data was
flushed to disk then it may just be a section of the file at the end that
has holes.

There are existing fixes to help alleviate this problem, particularly in
the case where a file has been truncated, that force cached data to be
flushed to disk when the file is closed.  If the system crashes while the
file(s) are still open then this flushing will never occur.

The fix that we have implemented is to introduce a second file size, called
the in-memory file size, that represents the current file size as viewed by
the user.  The existing file size, called the on-disk file size, is the one
that get's written to the filesystem log and we only update it when it is
safe to do so.  When we write to a file beyond eof we only update the in-
memory file size in the write operation.  Later when the I/O operation, that
flushes the cached data to disk completes, an I/O completion routine will
update the on-disk file size.  The on-disk file size will be updated to the
maximum offset of the I/O or to the value of the in-memory file size if the
I/O includes eof.
Merge of xfs-linux-melb:xfs-kern:28322a by kenmcd.

  Fix to prevent the notorious 'NULL files' problem after a crash.