Fix race in xfs_write() b/w dmapi callout and direct I/O checks.
In xfs_write() the iolock is dropped and reacquired in XFS_SEND_DATA() which
means that the file could change from not-cached to cached and we need to redo
the direct I/O checks.  We should also redo the direct I/O checks when the
file size changes regardless if O_APPEND is set or not.
Merge of xfs-linux-melb:xfs-kern:28440a by kenmcd.

  Fix race in xfs_write() b/w dmapi callout and direct I/O checks.