Sleeping with the ilock waiting for I/O completion is Bad.
Recent fixes to the filesystem freezing code introduced a vn_iowait
call in the middle of the sync code. Unfortunately, at the point
where this call was added we are holding the ilock. The ilock is
needed by I/O completion for unwritten extent conversion and now
updating the file size. Hence I/o cannot complete if we hol dthe
ilock while waiting for I/O completion.

Fix up the bug and clean the code up around it.
Merge of xfs-linux-melb:xfs-kern:28566a by kenmcd.

  Drop the ilock before calling vn_iowait() when doing a SYNC_IOWAIT
  sync operation. Make the code easier to understand as well.