Barriers need to be dynamically checked and switched off
If the underlying block device sudden stops supporting barriers,
we need to handle the -EOPNOTSUPP error in a sane manner rather
than shutting downteh filesystem. If we get this error, clear the
barrier flag, reissue the I/O, and tell the world bad things are
occurring.
Merge of xfs-linux-melb:xfs-kern:28568a by kenmcd.

  If the buffer gets a EOPNOTSUPP I/O error and it is a barrier
  write, clear the barrier and reissue the I/O.