Fix vmalloc leak on mount/unmount.
When setting the length of the iclogbuf to write out we should just
be changing the desired byte count rather completely reassociating
the buffer memory with the buffer. Reassociating the buffer memory
changes the apparent length of the buffer and hence when we free the
buffer, we don't free all the vmap()d space we originally allocated.
Merge of xfs-linux-melb:xfs-kern:28640a by kenmcd.

  Set the iclogbuf I/O size correctly rather than reassociating the
  memory which can lead to vmalloc leaks.