Make hole punching at EOF atomic.
If hole punching at EOF is done as two steps (i.e. truncate then extend)
the file is in a transient state between the two steps where an
application can see the incorrect file size. Punching a hole to
EOF needs to be treated in teh same way as all other hole punching
cases so that the file size is never seen to change.
Merge of xfs-linux-melb:xfs-kern:28641a by kenmcd.

  convert callers to new xfs_free_eofblocks() interface and allow it
  to be called with the iolock already held.