Write at EOF may not update filesize correctly.
The recent fix for preventing NULL files from being left around
does not update the file size corectly in all cases. The missing
case is a write extending the file that does not need to allocate
a block.

In that case we used a read mapping of the extent which forced
the use of the read I/O completion handler instead of the
write I/O completion handle. Hence the file size was not updated
on I/O completion.
Merge of xfs-linux-melb:xfs-kern:28657a by kenmcd.

  Ensure we use an I/O completion handler that can update the filesize
  for all types of write regardless of whether we need to do allocation
  or not.