xfs_bmapi fails to update the previous extent pointer
When processing multiple extent maps, xfs_bmapi needs to keep
track of the extent behind the one it is currently working on
to be able to trim extent ranges correctly. Failing to update
the previous pointer can result in corrupted extent lists
in memory and this will result in panics or assert failures.

Update the previous pointer correctly when we move to the
next extent to process.
Merge of xfs-linux-melb:xfs-kern:28773a by kenmcd.

  Update the previous extent pointer correctly in xfs_bmapi.