Fix remount,readonly path to flush everything correctly.
The remount readonly path can fail to writeback properly because we
still have active transactions after calling xfs_quiesce_fs().
Further investigation shows that this path is broken in the same
ways that the xfs freeze path was broken so fix it the same
way.
Merge of xfs-linux-melb:xfs-kern:28869a by kenmcd.

  Clean up data quiesce flags.