XFS should not be looking at filp reference counts
A check for file_count is always a bad idea.  Linux has the ->release
method to deal with cleanups on last close and ->flush is only for the
very rare case where we want to perform an operation on every drop of
a reference to a file struct.

This patch gets rid of vop_close and surrounding code in favour of
simply doing the page flushing from ->release.


Signed-off-by: Christoph Hellwig <hch@lst.de>
Merge of xfs-linux-melb:xfs-kern:28952a by kenmcd.

  Kill xfs_file_flush as xfs_file_release provides us with
  the required last-close callout.