Quota inode has no parent.
Avoid using a special "zero inode" as the parent of the quota inode
as this can confuse the filestreams code into thinking the quota inode
has a parent. We do not want the quota inode to follow filestreams allocation
rules, so pass a NULL as the parent inode and detect this condition
when doing stream associations.
Merge of xfs-linux-melb:xfs-kern:29098a by kenmcd.

  Check for inodes without parents and don't try to associate them
  or inherit from a non-existent parent inode.