Fix inode size update before data write in xfs_setattr
When changing the file size by a truncate() call, we log
the change in the inode size. However, we do not flush any outstanding
data that might not have been written to disk, thereby violating the
data/inode size update order. This can leave files full of NULLs
on crash.

Hence if we are truncating the file, flush any unwritten data that
may lie between the curret on disk inode size and the new inode size
that is being logged to ensure that ordering is preserved.
Merge of xfs-linux-melb:xfs-kern:29174a by kenmcd.

  Flush unwritten data between the current on disk inode size
  and the new inode size logged by xfs_setattr during a truncate
  operation to ensure the data is on disk before the inode size
  update.