On-demand reaping of the MRU cache
Instead of running the mru cache reaper all the time based on a
timeout, we should only run it when the cache has active objects.
This allows CPUs to sleep when there is no activity rather than
be woken repeatedly just to check if there is anything to do.
Merge of xfs-linux-melb:xfs-kern:29305a by kenmcd.

  Change the reaper to start when the first object is stored
  in the cache and stop when the last object is reaped.