cleanup defintions of BMBT_*BITLEN macros
The BMBT_*BITLEN are currently defined in a complicated way depending
on XFS_NATIVE_HOST.  But if all the macros are expanded they
(obviously) expand to the same value for both cases.

This patch defines the macros in the most simple way and updates the
comment describing them to remove outdated bits.


Signed-off-by: Christoph Hellwig <hch@lst.de>
Merge of xfs-linux-melb:xfs-kern:29320a by kenmcd.

  Remove the dependency of the BMBT_*BITLEN macros on XFS_NATIVE_HOST