choose single default logbuf count & size
Remove sizing of logbuf size & count based on physical memory;
this was never a very good gauge as it's looking at global
memory, but deciding on sizing per-filesystem; no account
is made of the total number of filesystems, for example.

For now just take the largest "default" case, as was set
for machines with >400MB - 8 x 32k buffers.  This can
always be tuned higher or lower with mount options
if necessary.  Removes one more user of xfs_physmem.

Signed-off-by: Eric Sandeen <sandeen@sandeen.net>
Merge of xfs-linux-melb:xfs-kern:29323a by kenmcd.

  Don't size logbuf size and count based on physical memory.