dinode endianess annotations
Biggest bit is duplicating the dinode structure so we have one annoted
for native endianess and one for disk endianess.  The other significant
change is that xfs_xlate_dinode_core is split into one helper per
direction to allow for proper annotations, everything else is trivial.

As a sidenode splitting out the incore dinode means we can move it into
xfs_inode.h in a later patch and severly improving on the include hell
in xfs.

Signed-off-by: Christoph Hellwig <hch@lst.de>
Merge of xfs-linux-melb:xfs-kern:29476a by kenmcd.

  add endian notations for the dinode.